% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acchist.r
\name{acchist}
\alias{acchist}
\title{Drawing a accummulate hist picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)}
\usage{
acchist(achist_data, individual, fill, value)
}
\arguments{
\item{achist_data}{the original data(panel or time series data) (dataframe)}

\item{individual}{the individual variable of different region or something else (factor)}

\item{fill}{Classification variables to indicate the type you care about of one of the individuals,like NewYork's 15-64/0-14/65+ age people. (factor)}

\item{value}{the value of the individual's index,like NewYork's birthrate.(numeric)}

\item{individual.name}{the names of the individual of different region or something else (charactor)}

\item{fill.name}{the names of  Classification variables to indicate the type you care about of one of the individuals,like NewYork's 15-64/0-14/65+ age people. (charactor)}

\item{bar_width}{the wide of the bar.(numeric)}

\item{axis.title.size}{the size of both x and y axis's title.(numeric)}

\item{axis.x.title}{the name of x axis.(charactor)}

\item{axis.y.title}{the name of y axis.(charactor)}

\item{axis.text.x.size}{the size of x axis's text.(numeric)}

\item{axis.text.y.size}{the size of y axis's text.(numeric)}

\item{legend.text.size}{the size of legend's text.(numeric)}

\item{legend.title.size}{the size of legend's title.(numeric)}

\item{Fon}{the Font of Picture}
}
\description{
Drawing a accummulate hist picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)
}
\examples{
library(chip)
data("acchist_test_data",package = "chip")  #load integration sample data
axis.x.title <- c("region")
axis.y.title <- c("ratio")
individual.name <- c("B","C","D","E","F")
fill.name <- c("academic","vocational","college","high","further","primary")
bar_width <- 0.3
axis.title.size <- 30
axis.text.x.size <- 30
axis.text.y.size <- 30
legend.text.size <- 30
legend.title.size <- 30
Fon <- 'sans'
acchist(acchist_test_data,acchist_test_data$region,acchist_test_data$edu,acchist_test_data$ratio)
}
