% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonbon.r
\name{bonbon}
\alias{bonbon}
\title{Drawing a bonbon picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)}
\usage{
bonbon(input_data, individual, value)
}
\arguments{
\item{input_data}{the original data(panel or time series data) (dataframe)}

\item{individual}{the individual variable of different region or something else (factor)}

\item{value}{the target value of the individual,like NewYork's birthrate.(numeric)}

\item{outstanding}{the individual you want to mark,if don't have any,please type:   outstanding <- c(" ")}

\item{outstanding.label}{the label of individual you wish to mark,if don't have any,please type:   outstanding.label <- c(" ")}

\item{outstanding.label.size}{size of the label of individual you wish to mark}

\item{outstanding.size}{wide of the outstanding bar}

\item{not.outstanding.size}{wide of the bar you don't want to mark}

\item{outstanding.head.size}{size of Round head(just for outstanding bar)}

\item{not.outstanding.head.size}{size of Round head(just for the bar you don't want to mark)}

\item{axis.title.x.size}{size of the title of x axis}

\item{axis.text.y.size}{size of the text of y axis}

\item{value.max}{max value of the sample you want to show}

\item{value.break}{the value you want to show in x axis}

\item{xlab.name}{the title of x axis}

\item{individual.name}{names of individual(drawing sequence is from bottom(first one) to top(last one))}
}
\description{
Drawing a bonbon picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)
}
\examples{
library(chip)
data("bonbon_test_data",package = "chip")  #load integration sample data
outstanding <- c(" ") or like this:
outstanding <- c("wholesaleretail","Leasing and business services","realestate")
outstanding.label <- c("24.75\%","28.51\%","5.16\%") #should be adjust with outstanding鎴笺埂鎸笺工outstanding.label
outstanding.label.size <- 7
outstanding.size <- 4
not.outstanding.size <- 2.3
outstanding.head.size <- 7
not.outstanding.head.size <- 5
axis.title.x.size <- 30
axis.text.y.size <- 30
value.max <- 100
value.break <- c(0,value.max\%/\%3,78,100)
xlab.name <- c("ratio of service industry(\%)")
individual.name <- c("Leasing and business services","wholesaleretail","scientificresearch","Accommodation and Catering","Informationtransmission","culture","socialwork","publicfacilities","residentsservice","financial","education","transportation","publicmanagement","realestate")
bonbon(bonbon_test_data,bonbon_test_data$industry,bonbon_test_data$index)
}
