% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxeasy.r
\name{boxeasy}
\alias{boxeasy}
\title{Drawing a box line picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)}
\usage{
boxeasy(input_data, individual, value)
}
\arguments{
\item{input_data}{the original data(panel or time series data) (dataframe)}

\item{individual}{the classification of value you want to show distribution,like year,region.}

\item{ylab.name}{y axis title}

\item{xlab.name}{x axis title}

\item{axis.title.x.size}{size of x axis title}

\item{axis.title.y.size}{size of y axis title}

\item{axis.text.x.size}{size of x axis text}

\item{axis.text.y.size}{size of y axis text}
}
\description{
Drawing a box line picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)
}
\examples{
library(chip)
data("box_test_data",package = "chip")  #load integration sample data
ylab.name <- c("index")
xlab.name <- c("year")
axis.title.x.size <- 15
axis.title.y.size <- 15
axis.text.x.size <- 15
axis.text.y.size <- 15
boxeasy(box_test_data,box_test_data$year,box_test_data$index)
}
