% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densline.r
\name{densline}
\alias{densline}
\title{Drawing a Kernel density picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)}
\usage{
densline(input_data)
}
\arguments{
\item{input_data}{the original data(panel or time series data) (dataframe)}

\item{xscope}{scope of x axis's value}

\item{yscope}{scope of y axis's value}

\item{legend.name}{names of every Kernal density line(input from the first column to the last column)}
}
\description{
Drawing a Kernel density picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)
}
