% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyline.r
\name{easyline}
\alias{easyline}
\title{Drawing a line picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)}
\usage{
easyline(input_data, value, time)
}
\arguments{
\item{value}{the target value to show}

\item{time}{the sequence,like:1,2,3,4;or 1998,1999,2000.....}

\item{easyline}{the original data(panel or time series data) (dataframe)}

\item{ylab.name}{title of y axis}

\item{xlab.name}{title of x axis}

\item{axis.title.x.size}{size of title of x axis}

\item{axis.title.y.size}{size of title of y axis}

\item{axis.line.x.thickness}{thickness of x axis}

\item{axis.line.y.thickness}{thickness of y axis}

\item{axis.text.x.size}{size of x axis's text}

\item{axis.text.y.size}{size of y axis's text}

\item{x.start}{start value of x axis's value}

\item{x.end}{end value of x axis's value}

\item{x.interval}{interval of x axis's value}

\item{y.start}{start value of y axis's value}

\item{y.end}{end value of y axis's value}

\item{y.interval}{interval of y axis's value}
}
\description{
Drawing a line picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)
}
\examples{
library(chip)
data("easyline_test_data",package = "chip")  #load integration sample data
ylab.name <- "population"
xlab.name <- "age"
axis.title.x.size <- 25
axis.title.y.size <- 25
axis.line.x.thickness <- 0.5
axis.line.y.thickness <- 0.5
axis.text.x.size <- 20
axis.text.y.size <- 20
x.start <- 0
x.end <- 100
x.interval <- 10
y.start <- 0
y.end <- 21571
y.interval <- 5000
easyline(easyline_test_data,easyline_test_data$population,easyline_test_data$age)
}
