% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histeasy.r
\name{histeasy}
\alias{histeasy}
\title{Drawing a hist picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)}
\usage{
histeasy(data, individual, type, value)
}
\arguments{
\item{individual}{the individual variable of different region or something else (factor)}

\item{type}{Classification variables to indicate the type you care about of one of the individuals,like NewYork's  15-64/0-14/65+ age people. (factor)}

\item{value}{the value of the individual's index,like NewYork's birthrate.(numeric)}

\item{hist_data}{the original data(panel or time series data) (dataframe)}

\item{axisname.x}{name of x axis (character)}

\item{axisname.y}{name of y axis (character)}

\item{type.name}{name of type,like"girl_of_age15-64" (character)}

\item{individual.name}{name of individual,like "NewYork" (character)}

\item{legend.x}{distance of legend(horizontal) (numeric)}

\item{legend.y}{distance of legend(vertical) (numeric)}

\item{legend.x.distance}{distance of legend's graphic and legend's font(horizontal) (numeric)}

\item{x.size}{size of x axis's font (numeric)}

\item{y.size}{size of y axis's font (numeric)}

\item{lab.size}{size of the font of title of both x axis and y axis (numeric)}
}
\description{
Drawing a hist picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)
}
\examples{
library(chip)
data("hist_test_data",package = "chip")  #load integration sample data
x.size <- 2 
y.size <- 2 
lab.size <- 2 
legend.x <- 30 
legend.y <- 0.3 
legend.x.distance <- 0.1 
axisname.x <- c("region") 
axisname.y <- c("ratio") 
type.name = c("0-14","15-64","65+","immortality") 
individual.name=c("A","B","C","D","E") 
histeasy(hist_test_data,hist_test_data$region,hist_test_data$age,hist_test_data$ratio)
}
