% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linepoint.r
\name{linepoint}
\alias{linepoint}
\title{Drawing a line picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)}
\usage{
linepoint(input_data, individual, value, time)
}
\arguments{
\item{input_data}{the original data(panel or time series data) (dataframe)}

\item{individual}{the individual variable of different region or something else (factor)}

\item{value}{value of target variable,like GDP愼㸱愼㸲CPI愼㸱愼㸲income.}

\item{time}{sequence,like year.}

\item{start}{start value of time}

\item{end}{end value of time}

\item{interval}{interval of time}

\item{axis.text.x.size}{size of x axis's text}

\item{axis.text.y.size}{size of y axis's text}

\item{legend.title.size}{size of title of legend}

\item{legend.text.size}{size of text of legend}

\item{axis.title.x.size}{size of title of x axis}

\item{axis.title.y.size}{size of title of y axis}

\item{axis.line.x.size}{thickness of x axis line}

\item{axis.line.y.size}{thickness of y axis line}

\item{geom.point.size}{size of point on line}

\item{ylab.name}{title of y axis}

\item{xlab.name}{title of x axis}

\item{individual.name}{name of individual}

\item{legend.name}{title of legend}

\item{Fon}{the Font of Picture}
}
\description{
Drawing a line picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)
}
\examples{
library(chip)
data("line_test_data",package = "chip")  #load integration sample data
start <- 2009
end <- 2019
interval <- 2
axis.text.x.size <- 30
axis.text.y.size <- 30
legend.title.size <- 30
legend.text.size <- 30
axis.title.x.size <- 30
axis.title.y.size <- 30
axis.line.x.size <- 0.5
axis.line.y.size <- 0.5
geom.point.size <- 7
ylab.name <- c("index")
xlab.name <- c("year")
individual.name <- c("A","B","C","D","E")
legend.name <- c("region")
Fon <- 'sans'
linepoint(line_test_data,line_test_data$region,line_test_data$index,line_test_data$year)
}
