% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramid.r
\name{pyramid}
\alias{pyramid}
\title{Drawing a pyramid picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)}
\usage{
pyramid(data, left, right, groupID, groupName)
}
\arguments{
\item{data}{the original data(panel or time series data) (dataframe)}

\item{left}{variable which you want to put in leftside}

\item{right}{variable which you want to put in rightside}

\item{groupID}{squence of group}

\item{groupName}{name of each group}

\item{v_max}{the max value of all variable}

\item{middle.size}{size of text in middle}

\item{left.axis.title.x.size}{size of left of x axis title}

\item{left.axis.text.x.size}{size of left of x axis text}

\item{right.axis.title.x.size}{size of right of x axis title}

\item{right.axis.text.x.size}{size of right of x axis text}

\item{right.title}{right title of x axis}

\item{left.title}{left title of x axis}

\item{Fon}{the Font of Picture}
}
\description{
Drawing a pyramid picture with Printing style
Any BUG You found,please send email to 1270950448@qq.com(DU Yunhan,from Southwestern University of Finance & Economics)
}
\examples{
data("pyramid_test_data",package = "chip")  #load integration sample data
v_max <- 4400
middle.size <- 6.6
left.axis.title.x.size <- 30
left.axis.text.x.size <- 30
right.axis.title.x.size <- 30
right.axis.text.x.size <- 30
Fon <- 'sans'
pyramid(pyramid_test_data,pyramid_test_data$male,pyramid_test_data$female,pyramid_test_data$group,pyramid_test_data$group_name)
}
